asdasdasdkjbvh2134

+++ src/main/java/ru/tuanviet/javabox/Benchmark.java
@@ -6,5 +6,5 @@ import java.lang.annotation.*;
 @Retention(value= RetentionPolicy.RUNTIME)
 public @interface Benchmark {
     int repeats();
-    int timeout();
+    long timeout();
 }
